radio: radio-tavarua: Re-org code for workqueue
When fm task is killed from task manager rpc failure is seen.
<3>[   72.817162] pmapp_rpc_req_reply: rpc failed! len = -512
<3>[   72.817173] fm_radio_shutdown: voting off failed with :(-512)
RPC failure is because data to be sent from apps, was on behalf of the
current process, but as the process is killed the current->pid given to
rpc is a non-existing pid.
To resolve the issue a valid pid is needed, which can be done by
queueing a work in existing workqueue and wait for completion.

Change-Id: I8181d0a398d79b493b92c899f068f008a8ff2d0f
Signed-off-by: Taniya Das <tdas@codeaurora.org>