Revert "radio-tavarua: Configure search direction as passed by FM APP"
As per V4L2 framework APP should pass the parameter as '1' to
search in forward direction and '0' to search in downward
direction. Reverting the change that was uploaded as it configured
the search direction directly that it received it from the APP.
APP should pass '1' for SEEK UP and '0' for SEEK DOWN . The driver
should check for the direction before configuring. To the FM SOC,
'0' means SEEK-UP and '1' means SEEK-DOWN. If APP passes '1' driver
understands that user is requesting SEEK-UP operation and will
configure the direction as '0'.

This reverts commit 75b90b2359fe55e9e98906fbc088a32fe5537675

Change-Id: I46d50994ec7eb52db389dc1b02e9a78600dbb5e1
Signed-off-by: Anantha Krishnan <ananthk@codeaurora.org>