msm: board8660: Allocate proper size to MM_FW heap
With CP-2.0 feature, mm_fw heap converted to hole of size
128 KB which lead to fail loading normal playback firmware.
So, increase the heap size to accommodate normal playback
firmware. Hole is now just a void space which won't be
visible to kernel. So, by this change, mm_fw heap is
overlapping with the secure firmware loading region.

Change-Id: I605fc748ea05602011de4f4592ca432bf627f531
Signed-off-by: Chintan Pandya <cpandya@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/board-msm8x60.c