gpu: ion: Add chunk heap
This patch adds support for a chunk heap that allows for buffers that are
made up of a list of fixed size chunks taken from a carveout.  Chunk sizes
are configured when the heaps are created by passing the chunk size in the
priv field of the heap platform data.

Change-Id: I188b96f0e1cb40240241984cfd3647817fec28e2
Signed-off-by: Rebecca Schultz Zavin <rebecca@android.com>
Git-commit: 2faceab86ef57bc60e8629bbbea02c057b4fd4c5
Git-repo: https://android.googlesource.com/kernel/common
[lauraa@codeaurora.org: Context differences in Makefile and ion.h due
to additional heaps added. Change sync API for newer kernel.
Changed print formats to actually be correct]
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>

Conflicts:
	include/linux/ion.h