gpu: ion: make cached Iommu allocations multiples of PAGE_SIZE
In future changes to Ion, we will fault in the pages one at a time
using vm_insert_page for cached allocations. There appears to be a
limitation on vm_insert_page that the pages must be of size
PAGE_SIZE. Make sure that we only allocate pages of size PAGE_SIZE for
cached allocations.

Change-Id: I2243e73c4f50b90c9345c25b9e4bc30ccfa7e198
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>