gpu: ion: don't fault in individual pages for the CP heap
The CP heap has some additional requirements regarding user mappings
that can't be satisfied with the current model of faulting in
individual pages. To fix this, use the CP heap's existing .map_user
callback to insert all of the pages into the VMA at mmap time, rather
than faulting them in on demand. This is the way it was always done
before the individual faulting scheme was introduced.

Change-Id: I9f1bbb2a0e65602bb74ab01d38e6e88d57327d9c
Signed-off-by: Mitchel Humpherys <mitchelh@codeaurora.org>